Extract Room database to separate module
