+++ core/src/main/kotlin/com/github/kr328/clash/core/model/Profile.kt
@@ -1,4 +1,4 @@
-package com.github.kr328.clash.service.model
+package com.github.kr328.clash.core.model
 
 import android.os.Parcelable
 import kotlin.uuid.Uuid
