+++ glue/build.gradle.kts
@@ -6,6 +6,7 @@ dependencies {
   api(projects.core)
   api(projects.core.model)
   api(projects.service)
+  api(projects.service.database)
   api(projects.common)
 
   implementation(libs.kotlin.coroutine.android)
