+++ service/build.gradle.kts
@@ -7,6 +7,7 @@ plugins {
 
 dependencies {
   implementation(projects.core.model)
+  implementation(projects.service.database)
 
   implementation(projects.core)
   implementation(projects.common)
@@ -14,10 +15,8 @@ dependencies {
   implementation(libs.kotlin.coroutine.android)
   implementation(libs.kotlin.serialization.json)
   implementation(libs.androidx.core)
-  implementation(libs.androidx.room.runtime)
   implementation(libs.kaidl.runtime)
   implementation(libs.rikkax.multiprocess)
 
   ksp(libs.kaidl.compiler)
-  ksp(libs.androidx.room.compiler)
 }
