+++ service/database/build.gradle.kts
@@ -1,0 +1,14 @@
+plugins {
+  alias(libs.plugins.android.library)
+  alias(libs.plugins.ksp)
+}
+
+dependencies {
+  implementation(projects.core)
+  implementation(projects.common)
+
+  implementation(libs.kotlin.coroutine.android)
+  implementation(libs.androidx.room.runtime)
+
+  ksp(libs.androidx.room.compiler)
+}
