+++ service/database/src/main/kotlin/com/github/kr328/clash/service/data/Imported.kt
@@ -2,7 +2,7 @@ package com.github.kr328.clash.service.data
 
 import androidx.room3.Entity
 import androidx.room3.PrimaryKey
-import com.github.kr328.clash.service.model.Profile
+import com.github.kr328.clash.core.model.Profile
 import kotlin.uuid.Uuid
 
 @Entity(tableName = "imported")
