+++ service/src/main/kotlin/com/github/kr328/clash/service/ProfileManager.kt
@@ -1,11 +1,11 @@
 package com.github.kr328.clash.service
 
 import android.content.Context
+import com.github.kr328.clash.core.model.Profile
 import com.github.kr328.clash.service.data.Database
 import com.github.kr328.clash.service.data.ImportedDao
 import com.github.kr328.clash.service.data.Pending
 import com.github.kr328.clash.service.data.PendingDao
-import com.github.kr328.clash.service.model.Profile
 import com.github.kr328.clash.service.remote.IFetchObserver
 import com.github.kr328.clash.service.remote.IProfileManager
 import com.github.kr328.clash.service.store.ServiceStore
