+++ service/src/main/kotlin/com/github/kr328/clash/service/ProfileReceiver.kt
@@ -11,9 +11,9 @@ import com.github.kr328.clash.common.compat.pendingIntentFlags
 import com.github.kr328.clash.common.constants.Intents
 import com.github.kr328.clash.common.util.componentName
 import com.github.kr328.clash.common.util.setUUID
+import com.github.kr328.clash.core.model.Profile
 import com.github.kr328.clash.service.data.Imported
 import com.github.kr328.clash.service.data.ImportedDao
-import com.github.kr328.clash.service.model.Profile
 import com.github.kr328.clash.service.util.importedDir
 import kotlin.time.Duration.Companion.minutes
 import kotlinx.coroutines.CoroutineScope
