+++ ui/profile/src/androidMain/kotlin/com/github/kr328/clash/profile/vm/NewProfileViewModel.kt
@@ -11,12 +11,12 @@ import com.github.kr328.clash.common.Res as CommonRes
 import com.github.kr328.clash.common.constants.Intents
 import com.github.kr328.clash.common.new_profile
 import com.github.kr328.clash.common.unknown
+import com.github.kr328.clash.core.model.Profile
 import com.github.kr328.clash.glue.util.withProfile
 import com.github.kr328.clash.profile.Res
 import com.github.kr328.clash.profile.import_from_qr_exception
 import com.github.kr328.clash.profile.import_from_qr_no_permission
 import com.github.kr328.clash.profile.model.ProfileProvider
-import com.github.kr328.clash.service.model.Profile
 import io.github.g00fy2.quickie.QRResult
 import kotlin.uuid.Uuid
 import kotlinx.coroutines.Dispatchers
