+++ ui/profile/src/androidMain/kotlin/com/github/kr328/clash/profile/vm/ProfilesViewModel.kt
@@ -5,13 +5,13 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.github.kr328.clash.common.Res as CommonRes
 import com.github.kr328.clash.common.unknown
+import com.github.kr328.clash.core.model.Profile
 import com.github.kr328.clash.glue.remote.Remote
 import com.github.kr328.clash.glue.util.withProfile
 import com.github.kr328.clash.profile.Res
 import com.github.kr328.clash.profile.active_unsaved_tips
 import com.github.kr328.clash.profile.toast_profile_updated_complete
 import com.github.kr328.clash.profile.toast_profile_updated_failed
-import com.github.kr328.clash.service.model.Profile
 import kotlin.time.Duration.Companion.minutes
 import kotlin.uuid.Uuid
 import kotlinx.coroutines.Dispatchers
